backend/requirements.txt
django==2.2.3
djangorestframework==3.10.3